Remove useless casts from missile-particle attributes.

+++ src/gui/windows/skilldialog.cpp
@@ -568,8 +568,8 @@ void SkillDialog::loadSkillData(XmlNodeConstPtr node,
         node, "missile-particle", "");
     missile.z = static_cast<float>(XML::getFloatProperty(
         node, "missile-z", 32.0f));
-    missile.lifeTime = static_cast<float>(XML::getProperty(
-        node, "missile-lifetime", 500));
+    missile.lifeTime = XML::getProperty(
+        node, "missile-lifetime", 500);
     missile.speed = static_cast<float>(XML::getFloatProperty(
         node, "missile-speed", 7.0f));
     missile.dieDistance = static_cast<float>(
@@ -580,8 +580,8 @@ void SkillDialog::loadSkillData(XmlNodeConstPtr node,
         node, "castingMissile-particle", "");
     castingMissile.z = static_cast<float>(XML::getFloatProperty(
         node, "castingMissile-z", 32.0f));
-    castingMissile.lifeTime = static_cast<float>(XML::getProperty(
-        node, "castingMissile-lifetime", 500));
+    castingMissile.lifeTime = XML::getProperty(
+        node, "castingMissile-lifetime", 500);
     castingMissile.speed = static_cast<float>(XML::getFloatProperty(
         node, "castingMissile-speed", 7.0f));
     castingMissile.dieDistance = static_cast<float>(
